Add animations back and create a page handler
I didn't want to rely on using IDs to identify different elements so I made page-handler to handle modifications to the content div. For example it's used for the buttons in the header to change which page is shown. I also added the animations back to the buttons in the header.