Change how page-handler changes between pages
I noticed that I had many functions that basically did the same thing, so I tried to simplify the code a bit. Now theres a changePage function that get's called. It will set the state, clear the page, and then call the showPage function which determines which page to show. changePage is also used by the goBack function when the user presses the back button. Now instead of importing different functions to the other modules, they can all import StateName and changePage to do the same thing.